requirements.txt
webdriver-manager
selenium==4.8.0
pandas
openpyxl
python-dotenv